(try to) fetch enrollments
This code fetches and deserializes enrollments, but the API call for it
fails with a 500 from the server. Nothing in the documentation indicates
I am making the call incorrectly. But at least we have some prior art
now on how to get a collection of futures and eventually get the
resulting value we want.